+++ source/CHANGELOG.md
@@ -4,6 +4,7 @@ Andreas Abel <andreas.abel@gu.se>
 
 * BNFC now uniformly signs generated files with its version number [#373]
 * C++: fixed parser regression in 2.9.2: missing `#include <algorithm>` [#377]
+* Ocaml: lex CR as whitespace [see also #376]
 
 # 2.9.2